requirements.txt
# if you want to upgrade these packages be careful
# it took me a lot of time to choose the correct version to make
# those models works 
# tested on pip 25.2
pillow>=9.0.0
rasterio>=1.3a3
numpy==1.26.4
requests>=2.25.0
pyyaml>=5.3.1
ultralytics>=8.3.184
matplotlib>=3.5.0
scikit-learn>=1.0.0
opencv-python>=4.5.0
torch>=1.11.0
geopandas>=0.13.0
shapely>=2.0.0
detectron2 @ git+https://github.com/facebookresearch/detectron2.git
detectree2 @ git+https://github.com/PatBall1/detectree2.git